server: Renamed socialMedia to social-media
