+++ server/api/root.ts
@@ -5,7 +5,7 @@ import {experienceRouter} from "./routers/experience";
 import {imageRouter} from "./routers/image";
 import {projectRouter} from "./routers/project";
 import {snippetRouter} from "./routers/snippet";
-import {socialMediaRouter} from "./routers/socialMedia";
+import {socialMediaRouter} from "./routers/social-media";
 
 /**
  * This is the primary router for your server.
